% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein1D.R
\name{bernstein1D}
\alias{bernstein1D}
\title{bernstein1D}
\usage{
bernstein1D(x, y)
}
\arguments{
\item{x}{Numeric scalar in the closed interval [0,1]. The value at which the function will be approximated.}

\item{y}{Numeric vector. The function evaluated at a regular partition of the closed interval [0,1].}
}
\value{
Numeric scalar, the value of the Bernstein-Bezier curve.
}
\description{
Univariate Bernstein-Bezier function.
}
\examples{
# Example 1:
ex <- seq(0,1, length.out=5)
ey <- sin(2*pi*ex); plot(ex, ey)
exB <- seq(0,1, by = 0.01)
eyB <- sapply(exB, bernstein1D, y = ey)
points(exB,eyB, pch = "+")

# Example 2 (Bernstein-Kantorovich)
eyE <- c(ey[1], (ey[-1] + head(ey, -1))/2, tail(ey, 1))
eyEB <- sapply(exB, bernstein1D, y = eyE)
}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
