% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalFunc2D.R
\name{evalFunc2D}
\alias{evalFunc2D}
\title{evaluates a bivariate function in a list with the function name and named parameters.}
\usage{
evalFunc2D(f, x, y, p = 1)
}
\arguments{
\item{f}{A list with the function name and named parameters. See examples.}

\item{x}{Independent variable.}

\item{y}{Vector where \code{f} will be evaluated. If \code{f} is not a list, \code{x, y} is not used.}

\item{p}{Non-negative integer. The index where the name (\code{f}) is.}
}
\value{
If is a list, the evaluation of \code{f} at points \code{x,y}. If \code{f} is not a list, the result is the input \code{f}.
}
\description{
evaluates a bivariate function in a list with the function name and named parameters.
}
\details{
x and y go to the first and second input parameter of the function \code{f}.
}
\examples{
ss <- function(x, y, m = 5) {m * x + y}
evalFunc2D(f = list(f = "outer", FUN = "ss", m = 10), x = 1:4, y = 6:7)
}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
