% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fn.inv.Bernshtein.R
\name{Fn.inv.Bernshtein}
\alias{Fn.inv.Bernshtein}
\title{Bersntein polynomial fitting to a quasi-inverse.}
\usage{
Fn.inv.Bernshtein(u, valores.emp)
}
\arguments{
\item{u}{A point where the fitted Bernstein polynomial is to be evaluated at. \eqn{u \in [0,1]}}

\item{valores.emp}{Observed values.}
}
\description{
Bersntein polynomial fitting to a quasi-inverse.
}
\details{
See also equation 4 in Hernandez-Maldonado, Diaz-Viera and Erdely, 2012,
A joint stochastic simulation method using the Bernstein copula as a flexible...
This is the same function as lmomco:::dat2bernqua
}
\references{
Munoz-Perez and Fernandez-Palacin, 1987. Bernstein-Kantorovich polynomial
}
