% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupDuplicatedUtil.R
\name{groupDuplicatedUtil}
\alias{groupDuplicatedUtil}
\title{Manage/remove duplicated values in a table}
\usage{
groupDuplicatedUtil(x, col = 1, fun = "median")
}
\arguments{
\item{x}{A numeric table (data.frame or array).}

\item{col}{The column to be checked for duplicated values.}

\item{fun}{The function to process duplicated values.}
}
\value{
x with duplicated values processed.
}
\description{
Manage/remove duplicated values in a table
}
\examples{
ex <- data.frame(x=c(1,-2,3,4,-2,4), y = c(10,11,10,12,13,11))
ex1 = groupDuplicatedUtil(ex); print(ex1)
groupDuplicatedUtil(ex1, col = 2)
groupDuplicatedUtil(ex, col = 2)
}
