% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet.R
\name{sheet}
\alias{sheet}
\title{Computes a 2D function values over a given area.}
\usage{
sheet(f, lims, n = 101, plot = TRUE, ...)
}
\arguments{
\item{f}{A list with {name}, other named arguments to \code{f}. See example below.}

\item{lims}{a 2 x 2 array or data frame. the first row are the limits in the x axis, and the row 2 are the limits of the y axis.}

\item{n}{a single scalar or a 2-elements array with the number of values in each direction.}

\item{plot}{logical. Should display a plot?}

\item{...}{further arguments passed to the function \link[graphics]{image.default}.}
}
\value{
An array of the values computed.
}
\description{
Computes a 2D function values over a given area.
}
\examples{
# Example 1
ss <- function(x, y, m = 5) {m * x + y}
ex0=1; exf=4; ey0=6; eyf= 7
elims <- rbind(x = c(ex0, exf), y = c(ey0, eyf)); print(elims)
ez <- sheet(f=list(name = ss, m=10),
      lims= elims,
      n = c(exf-ex0 + 1, eyf-ey0 + 1))
print(ez)

# Example 2
ez <- sheet(f=list(name=ss), lims = elims)
}
\seealso{
\code{sheet} is the 2D analog of \link[graphics]{curve}.
}
