% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xySim.R
\name{xySim}
\alias{xySim}
\title{Generate simulations from bivariate dataset}
\usage{
xySim(n = 100, xy, diffEC = forwardDifference(empiricalCDF2Dcounts(xy)),
  xcond = NULL)
}
\arguments{
\item{n}{An integer specifying the number of simulations required.}

\item{xy}{Numeric 2-columns table (data.frame, matrix, cbind) of real dataset. The first column is the independent variable while the second is the dependent variable.}

\item{diffEC}{Forward difference with respect of $x$ of the empirical copula matrix.
Maybe computed with \link[empiricalDistribution]{forwardDifference} of \link[empiricalDistribution]{empiricalCDF2Dcounts}. \link[empiricalDistribution]{empiricalCDF2Dcounts}.}

\item{xcond}{A numeric vector being the 1st column of xy (xy[,1]). Must be conditional to the independent variable (\code{x})? Default to \code{NULL}.}
}
\value{
A list with
 xy_sim: 2-columns matrix of simulated observations \code{x,y}.
 uv_sim: 2-columns matrix of simulated pseudo observations \code{u,v}.
}
\description{
Generate simulations from bivariate datasets.
}
\examples{
# Example 1
data("PHIV_K"); plot(PHIV_K)
eSim <- xySim(n = 200, xy = PHIV_K)
plot(eSim$xy_sim)
}
