% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernstein1D.R
\name{bernstein1D}
\alias{bernstein1D}
\title{bernstein1D}
\usage{
bernstein1D(x, y)
}
\arguments{
\item{x}{Numeric scalar in the closed interval [0,1]. The value at which the function will be approximated.}

\item{y}{Numeric vector. The function evaluated at a regular partition of the closed interval [0,1].}
}
\value{
Numeric scalar, the value of the Bernstein-Bezier curve.
}
\description{
Univariate Bernstein-Bezier function.
}
\examples{
ex <- seq(0,1, length.out=5)
eF <- sin(2*pi*ex); plot(ex, eF)
exB <- seq(0,1, by = 0.01)
for(i in 1:length(exB)){
eFB <- bernstein1D(x=exB[i], y=eF);
points(exB[i], eFB, pch = '+')
}

}
\author{
Francisco Mendoza-Torres (\email{mentofran@gmail.com})
}
